msm: gsi: change some types in gsi_evt_ring_props
Change the ring_len field to have type u32, and the ring_base_addr
to have type dma_addr_t in the gsi_evt_ring_props structure.  This
is in preparation for the next patch.

+++ drivers/platform/msm/ipa_lite/gsi_dbg.c
@@ -37,7 +37,7 @@ static ssize_t gsi_dump_evt(struct file *file,
 	char *sptr, *token;
 	uint32_t val;
 	struct gsi_evt_ctx *ctx;
-	uint16_t i;
+	u32 i;
 
 	if (sizeof(dbg_buff) < count + 1)
 		return -EFAULT;
